Rename type to styleType in Button component
type as a prop name is contradictory to the html type which could be passed as an additional
prop that should be applied to the button. Renaming this prop from type to styleType
will allow type to be passed as an additional prop that can now be applied to the button itself.