Update Alert component
Fix issue with Icons not displaying properly:

The iconMap had 'exclamation-sign' spelled wrong for both warning and danger icons. This was causing the icons to render incorrectly.
Rename the iconMap values from 'exclaimation-sign' to 'exclamation-sign' in order to get them to render properly.

Update propTypes to be less restrictive:

Do not require additionalText prop or header prop.

Pass other props:

Allow any additional props to be passed to the outer div that wraps the entire Alert.