Update Alert Component
Add extraNotifcationText prop:

Takes in a string and adds a 'bell' icon and the provided text to the upper right corner of the alert. This allows for any potential notifications that need to be displayed along with the al
ert.

Update Alert styles:

- line up the body text with the leading header when there is an icon included
- added a 15px padding between the leading header and body text

Update Alert documentation:

Updated the Alert docs with the extraNotifcationText example.