Allow the default first option to be selected
Dropdowns are currently forcing selections to be made if any option besides the first option is selected. If a user wanted to select the first option to indicate no value, they would be unable to do so because it is disabled by default.

This adds a prop to control whether or not the default first option is disabled.