Remove genId from Textbox component
Provide an id prop that is optional.

If no ID prop is provided, use the input name and create an id for the textbox to be used in the htmlFor for the label and the id of the input. This ID is also used to generate an ID for the help content if needed.