Fix icons within Notification component
The icons in the Notification component were mapped to invalid entries in the icon-definitons.js file. This caused a different icon to be displayed, or no icon to be displayed at all when the includeIcon prop was set to true.

Changing the icon map to point to icons that are in the UX Guidelines fixes this problem.