Make theme props required
The current approach to not making theme required and providing a
default value of an empty object lead to false positives with TabPanel;
the prop types may have validated correctly, but the empty objects were
passed to the theme and incorrectly rendered the resulting tabs. We made
the prop types from a user perspective by not requiring the theme prop,
but in order to operate the affected components NEED a theme object
provided to them in some shape or form. For the time being making the
props required will avoid the aforementioned situation.