Update tab panels to give parent more control
Rather than control which tab is selected by ourselves, I moved the
logic into props which require the callback to change the prop.
The reason for the change is downstream when we really started using it,
the metadata around the content would change, but because we make a copy
of the elements they don't rerender correctly. This fixes things down
the line and makes the component more useful overall.