Fix: add babel-polyfill to head
For IE11 compatibility, there are some missing features that babel does
not support out of the box. babel-polyfill is a package that fills in
those gaps and allows common ES6 features to work even in IE11.

Add a script tag that points to babel-polyfill to the head of the template
in order to ensure that the polyfill is loaded before all components are
loaded.