Update: attempt to use id from props using useUniqueId
Downstream snapshots have a problem with fully generated unique identifiers. Use a passed in id prop if available for the unique identifier for any component that is currently using the useUniqueId hook.

Remove the useUniqueId from breadcrumb and use the index from map as the key. This is safe because there is no manipulation of the list being rendered.