Fix: correct css and size prop of FullColorIcon
The css prop previously normalized the size assignment to the container
width and height by suffixing the value with 'px' if the string was not
present. The logic from Icon was reused to patch this. The css rules
were also incorrectly formatted after being converted from a css prop to
a full styled component.